+++ dali/test/python/test_operator_readers_webdataset_big.py
@@ -1,0 +1,1 @@
+from test_operator_readers_webdataset_base import *
